Use package data to define a Seqinfo object

Use package data to define a Seqinfo object

defineSeqinfo(
  build = c("GRCh38", "GRCh37", "GRCm39", "GRCm38"),
  chr = TRUE,
  mito,
  ...
)

Arguments

build

The Genome build used

chr

logical(1) Include the prefix "chr"

mito

Specify M or MT to include the mitochondrial chromosome. Omitted by default

...

Not used

Value

A Seqinfo object

Details

This function will create a Seqinfo object from pre-defined data from the Genome Reference Consortium. Returned objects will always be restricted to assembled molecules only. Currently implemented genome builds represent the four most common builds for ChIP-Seq analysis

Examples

defineSeqinfo("GRCh37", TRUE)
#> Seqinfo object with 24 sequences from GRCh37 genome:
#>   seqnames seqlengths isCircular genome
#>   chr1      249250621      FALSE GRCh37
#>   chr2      243199373      FALSE GRCh37
#>   chr3      198022430      FALSE GRCh37
#>   chr4      191154276      FALSE GRCh37
#>   chr5      180915260      FALSE GRCh37
#>   ...             ...        ...    ...
#>   chr20      63025520      FALSE GRCh37
#>   chr21      48129895      FALSE GRCh37
#>   chr22      51304566      FALSE GRCh37
#>   chrX      155270560      FALSE GRCh37
#>   chrY       59373566      FALSE GRCh37
defineSeqinfo("GRCh37", FALSE, "MT")
#> Seqinfo object with 25 sequences (1 circular) from GRCh37 genome:
#>   seqnames seqlengths isCircular genome
#>   1         249250621      FALSE GRCh37
#>   2         243199373      FALSE GRCh37
#>   3         198022430      FALSE GRCh37
#>   4         191154276      FALSE GRCh37
#>   5         180915260      FALSE GRCh37
#>   ...             ...        ...    ...
#>   21         48129895      FALSE GRCh37
#>   22         51304566      FALSE GRCh37
#>   X         155270560      FALSE GRCh37
#>   Y          59373566      FALSE GRCh37
#>   MT            16571       TRUE GRCh37