Make a profile heatmap

Make a profile heatmap with optional summary panel at the top

.makeFinalProfileHeatmap(
  data,
  x = NULL,
  y = NULL,
  fill = NULL,
  colour = NULL,
  linetype = NULL,
  facet_x = NULL,
  facet_y = NULL,
  summary_fun = c("mean", "median", "min", "max", "none"),
  rel_height = 0.3,
  x_lab = NULL,
  y_lab = NULL,
  fill_lab = NULL,
  ...
)

Arguments

data
A data.frame or tibble in long form

x, y
The values mapped to the x & y axes

fill
The column used for heatmap colours

colour, linetype
Columns used for the summary plot in the top panel

facet_x, facet_y
Columns used to facet the plot along these axes

summary_fun
Function used to create the summary value at each position

rel_height
The relative height of the top panel

x_lab, y_lab, fill_lab
_labels added to x/y-axes & the fill legend

...
Passed to facet_grid

Value

A ggplot2 object

Details

The workhorse function for generating the final heatmap Expects a single data.frame in long form with requisite columns