Apply two filters to sliding windows

Apply two filters to counts generated using sliding windows

dualFilter(x, bg, ref, ...)

# S4 method for RangedSummarizedExperiment,RangedSummarizedExperiment,GRanges
dualFilter(x, bg, ref, q = 0.99, logCPM = TRUE, BPPARAM = bpparam())

Arguments

x

RangedSummarizedExperiment containing sample counts

bg

RangedSummarizedExperiment containing background/input counts

ref

GRanges object containing ranges where signal is expected

...

Not used

q

The upper percentile of the reference ranges expected to be returned when tuning the filtering criteria

logCPM

logical(1) Add a logCPM assay to the returned data

BPPARAM

Settings for running in parallel

Value

A RangedSummarizedExperiment which is a filtered subset of the original object. If requested the assay "logCPM" will be added (TRUE by default)

Details

This function will take sliding (or tiling) windows for it's input as a RangedSummarizedExperiment object. The dual strategy of applying filterWindowsControl and filterWindowsProportion will then be applied. A set of reference ranges for which signal is expected is used to refine the filtering criteria.

Cutoff values are found such that the q of the windows which overlap one of the reference ranges will be returned, along with any others which match the filtering criteria. Cutoff values for both criteria are added to the metadata element of the returned object