Datasets for an example region

Various example datasets for demonstrating analysis and visualisation strategies. Generation of all datasets is documented in system.file("script/ex_datasets.md", package = "extraChIPs")

ex_genes

Simple GRanges object with complete ranges for each gene

ex_trans

Exon & transcript level information prepared for plotting with Gviz or plotHFGC()

ex_prom

Regions defined as promoters

ex_hic

Example HiC interactions

data(ex_trans)

data(ex_genes)

data(ex_prom)

data(ex_hic)

Format

GRanges and GInteractions objects

All annotations are from GRCh37

An object of class GRanges of length 4.

An object of class GRanges of length 9.

An object of class GInteractions of length 1.

Examples

data(ex_trans)
ex_trans
#> GRanges object with 175 ranges and 6 metadata columns:
#>                   seqnames              ranges strand |     type
#>                      <Rle>           <IRanges>  <Rle> | <factor>
#>   ENST00000517259    chr10 103862019-103862089      + |     exon
#>   ENST00000413464    chr10 103892751-103892978      + |     exon
#>   ENST00000278070    chr10 103892797-103892978      + |     exon
#>   ENST00000413464    chr10 103897607-103897795      + |     exon
#>   ENST00000278070    chr10 103897607-103897795      + |     exon
#>               ...      ...                 ...    ... .      ...
#>   ENST00000361198    chr10 103871191-103871293      - |     exon
#>   ENST00000425280    chr10 103871191-103871293      - |     exon
#>   ENST00000361198    chr10 103874163-103874703      - |     exon
#>   ENST00000425280    chr10 103879843-103880042      - |     exon
#>   ENST00000673968    chr10 103879843-103880125      - |     exon
#>                              gene               exon      transcript
#>                       <character>        <character>     <character>
#>   ENST00000517259 ENSG00000253068  ENST00000517259_1 ENST00000517259
#>   ENST00000413464 ENSG00000148840  ENST00000413464_1 ENST00000413464
#>   ENST00000278070 ENSG00000148840  ENST00000278070_1 ENST00000278070
#>   ENST00000413464 ENSG00000148840  ENST00000413464_2 ENST00000413464
#>   ENST00000278070 ENSG00000148840  ENST00000278070_2 ENST00000278070
#>               ...             ...                ...             ...
#>   ENST00000361198 ENSG00000198728 ENST00000361198_10 ENST00000361198
#>   ENST00000425280 ENSG00000198728 ENST00000425280_10 ENST00000425280
#>   ENST00000361198 ENSG00000198728 ENST00000361198_11 ENST00000361198
#>   ENST00000425280 ENSG00000198728 ENST00000425280_11 ENST00000425280
#>   ENST00000673968 ENSG00000198728 ENST00000673968_11 ENST00000673968
#>                        symbol      status
#>                   <character> <character>
#>   ENST00000517259    SNORD112  Undetected
#>   ENST00000413464       PPRC1          Up
#>   ENST00000278070       PPRC1          Up
#>   ENST00000413464       PPRC1          Up
#>   ENST00000278070       PPRC1          Up
#>               ...         ...         ...
#>   ENST00000361198        LDB1   Unchanged
#>   ENST00000425280        LDB1   Unchanged
#>   ENST00000361198        LDB1   Unchanged
#>   ENST00000425280        LDB1   Unchanged
#>   ENST00000673968        LDB1   Unchanged
#>   -------
#>   seqinfo: 24 sequences from GRCh37 genome