Genes in an example region

Genes in an example region formatted for Gviz

ex_genes

Format

Gene structure as a GRanges object, in the correct structure for plotting with plotHFGC. This is a GRangesList with two elements: 'Unchanged' and 'Up' representing genes that were considered as unchanged or upregulated in a separate RNA-seq analysis. Each element is pre-formatted for compatability with the package Gviz and has the columns

type

The type of feature. All here are exons

gene

The Ensembl gene ID

exon

An arbitrary exon ID

transcript

The Ensembl transcript ID

symbol

The gene nme (or symbol)

All annotations are from GRCh37

Examples

data(ex_genes)
ex_genes
#> GRangesList object of length 2:
#> $Unchanged
#> GRanges object with 36 ranges and 5 metadata columns:
#>                   seqnames              ranges strand |     type
#>                      <Rle>           <IRanges>  <Rle> | <factor>
#>   ENST00000673968    chr10 103866246-103868080      - |     exon
#>   ENST00000361198    chr10 103867317-103868080      - |     exon
#>   ENST00000425280    chr10 103867317-103868080      - |     exon
#>   ENST00000673968    chr10 103868786-103868934      - |     exon
#>   ENST00000361198    chr10 103868786-103868934      - |     exon
#>               ...      ...                 ...    ... .      ...
#>   ENST00000361198    chr10 103871191-103871293      - |     exon
#>   ENST00000425280    chr10 103871191-103871293      - |     exon
#>   ENST00000361198    chr10 103874163-103874703      - |     exon
#>   ENST00000425280    chr10 103879843-103880042      - |     exon
#>   ENST00000673968    chr10 103879843-103880125      - |     exon
#>                              gene               exon      transcript
#>                       <character>        <character>     <character>
#>   ENST00000673968 ENSG00000198728  ENST00000673968_1 ENST00000673968
#>   ENST00000361198 ENSG00000198728  ENST00000361198_1 ENST00000361198
#>   ENST00000425280 ENSG00000198728  ENST00000425280_1 ENST00000425280
#>   ENST00000673968 ENSG00000198728  ENST00000673968_2 ENST00000673968
#>   ENST00000361198 ENSG00000198728  ENST00000361198_2 ENST00000361198
#>               ...             ...                ...             ...
#>   ENST00000361198 ENSG00000198728 ENST00000361198_10 ENST00000361198
#>   ENST00000425280 ENSG00000198728 ENST00000425280_10 ENST00000425280
#>   ENST00000361198 ENSG00000198728 ENST00000361198_11 ENST00000361198
#>   ENST00000425280 ENSG00000198728 ENST00000425280_11 ENST00000425280
#>   ENST00000673968 ENSG00000198728 ENST00000673968_11 ENST00000673968
#>                        symbol
#>                   <character>
#>   ENST00000673968        LDB1
#>   ENST00000361198        LDB1
#>   ENST00000425280        LDB1
#>   ENST00000673968        LDB1
#>   ENST00000361198        LDB1
#>               ...         ...
#>   ENST00000361198        LDB1
#>   ENST00000425280        LDB1
#>   ENST00000361198        LDB1
#>   ENST00000425280        LDB1
#>   ENST00000673968        LDB1
#>   -------
#>   seqinfo: 24 sequences from GRCh37 genome
#> 
#> $Up
#> GRanges object with 48 ranges and 5 metadata columns:
#>                   seqnames              ranges strand |     type
#>                      <Rle>           <IRanges>  <Rle> | <factor>
#>   ENST00000413464    chr10 103892751-103892978      + |     exon
#>   ENST00000278070    chr10 103892797-103892978      + |     exon
#>   ENST00000413464    chr10 103897607-103897795      + |     exon
#>   ENST00000278070    chr10 103897607-103897795      + |     exon
#>   ENST00000413464    chr10 103898376-103898522      + |     exon
#>               ...      ...                 ...    ... .      ...
#>   ENST00000278070    chr10 103908935-103909086      + |     exon
#>   ENST00000370012    chr10 103908935-103909086      + |     exon
#>   ENST00000370012    chr10 103909683-103909937      + |     exon
#>   ENST00000413464    chr10 103909683-103910088      + |     exon
#>   ENST00000278070    chr10 103909683-103910090      + |     exon
#>                              gene               exon      transcript
#>                       <character>        <character>     <character>
#>   ENST00000413464 ENSG00000148840  ENST00000413464_1 ENST00000413464
#>   ENST00000278070 ENSG00000148840  ENST00000278070_1 ENST00000278070
#>   ENST00000413464 ENSG00000148840  ENST00000413464_2 ENST00000413464
#>   ENST00000278070 ENSG00000148840  ENST00000278070_2 ENST00000278070
#>   ENST00000413464 ENSG00000148840  ENST00000413464_3 ENST00000413464
#>               ...             ...                ...             ...
#>   ENST00000278070 ENSG00000148840 ENST00000278070_13 ENST00000278070
#>   ENST00000370012 ENSG00000148840  ENST00000370012_9 ENST00000370012
#>   ENST00000370012 ENSG00000148840 ENST00000370012_10 ENST00000370012
#>   ENST00000413464 ENSG00000148840 ENST00000413464_12 ENST00000413464
#>   ENST00000278070 ENSG00000148840 ENST00000278070_14 ENST00000278070
#>                        symbol
#>                   <character>
#>   ENST00000413464       PPRC1
#>   ENST00000278070       PPRC1
#>   ENST00000413464       PPRC1
#>   ENST00000278070       PPRC1
#>   ENST00000413464       PPRC1
#>               ...         ...
#>   ENST00000278070       PPRC1
#>   ENST00000370012       PPRC1
#>   ENST00000370012       PPRC1
#>   ENST00000413464       PPRC1
#>   ENST00000278070       PPRC1
#>   -------
#>   seqinfo: 24 sequences from GRCh37 genome
#>