Import peaks

Import peaks in narrowPeak or broadPeak format

importPeaks(
  x,
  type = c("narrow", "broad"),
  blacklist,
  seqinfo,
  pruning.mode = c("coarse", "error"),
  sort = TRUE,
  setNames = TRUE,
  ...
)

Arguments

x

One or more files to be imported. All files must be of the same type, i.e. narrow or broad

type

The type of peaks to be imported

blacklist

A set of ranges to be excluded

seqinfo

A seqinfo object to be applied to the GRanges objects

pruning.mode

How to handle conflicts if supplying a seqinfo object. Defaults to pruning.mode = "coarse". Only "coarse" and "error" are implemented. See seqinfo.

sort

logical. Should the ranges be sorted during import

setNames

logical Set basename(x) as the name

...

passed to sort

Value

A GRangesList

Details

Peaks are imported from either narrowPeak or broadPeak format as GenomicRanges objects.

Examples

fl <- system.file(
    c("extdata/ER_1.narrowPeak", "extdata/ER_2.narrowPeak"),
    package = "extraChIPs"
)
peaks <- importPeaks(fl)
peaks
#> GRangesList object of length 2:
#> $ER_1.narrowPeak
#> GRanges object with 10 ranges and 5 metadata columns:
#>                                     seqnames          ranges strand |     score
#>                                        <Rle>       <IRanges>  <Rle> | <numeric>
#>    T47D_ER_E2_1_mapped_reads_peak_1     chr1   856458-856640      * |        56
#>    T47D_ER_E2_1_mapped_reads_peak_2     chr1 1008550-1010075      * |      6357
#>    T47D_ER_E2_1_mapped_reads_peak_3     chr1 1014771-1016015      * |       951
#>    T47D_ER_E2_1_mapped_reads_peak_4     chr1 1051383-1051896      * |       655
#>    T47D_ER_E2_1_mapped_reads_peak_5     chr1 1342664-1342846      * |        66
#>    T47D_ER_E2_1_mapped_reads_peak_6     chr1 1368480-1368979      * |       225
#>    T47D_ER_E2_1_mapped_reads_peak_7     chr1 1608098-1608345      * |       723
#>    T47D_ER_E2_1_mapped_reads_peak_8     chr1 1790733-1790975      * |        77
#>    T47D_ER_E2_1_mapped_reads_peak_9     chr1 1878927-1879257      * |       338
#>   T47D_ER_E2_1_mapped_reads_peak_10     chr1 1900588-1900902      * |       225
#>                                     signalValue    pValue    qValue      peak
#>                                       <numeric> <numeric> <numeric> <numeric>
#>    T47D_ER_E2_1_mapped_reads_peak_1     5.65444   8.52782   5.65761        91
#>    T47D_ER_E2_1_mapped_reads_peak_2   149.37155 641.42468 635.74786       662
#>    T47D_ER_E2_1_mapped_reads_peak_3    33.45546  98.85455  95.12226       210
#>    T47D_ER_E2_1_mapped_reads_peak_4    25.44500  69.13743  65.55376       182
#>    T47D_ER_E2_1_mapped_reads_peak_5     6.12565   9.59455   6.68896       108
#>    T47D_ER_E2_1_mapped_reads_peak_6    12.25129  25.77251  22.54826       133
#>    T47D_ER_E2_1_mapped_reads_peak_7    27.32981  75.94558  72.32353       149
#>    T47D_ER_E2_1_mapped_reads_peak_8     6.59685  10.69307   7.75227       115
#>    T47D_ER_E2_1_mapped_reads_peak_9    16.02092  37.24069  33.88997       160
#>   T47D_ER_E2_1_mapped_reads_peak_10    12.25129  25.77251  22.54826       169
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $ER_2.narrowPeak
#> GRanges object with 10 ranges and 5 metadata columns:
#>                                     seqnames          ranges strand |     score
#>                                        <Rle>       <IRanges>  <Rle> | <numeric>
#>    T47D_ER_E2_2_mapped_reads_peak_1     chr1   868541-868839      * |        97
#>    T47D_ER_E2_2_mapped_reads_peak_2     chr1 1008607-1010057      * |      6050
#>    T47D_ER_E2_2_mapped_reads_peak_3     chr1 1014770-1015980      * |       783
#>    T47D_ER_E2_2_mapped_reads_peak_4     chr1 1051307-1051918      * |       959
#>    T47D_ER_E2_2_mapped_reads_peak_5     chr1 1296000-1296181      * |        66
#>    T47D_ER_E2_2_mapped_reads_peak_6     chr1 1310420-1310703      * |        55
#>    T47D_ER_E2_2_mapped_reads_peak_7     chr1 1368372-1369009      * |       291
#>    T47D_ER_E2_2_mapped_reads_peak_8     chr1 1608102-1608346      * |       648
#>    T47D_ER_E2_2_mapped_reads_peak_9     chr1 1690460-1690641      * |        35
#>   T47D_ER_E2_2_mapped_reads_peak_10     chr1 1878958-1879250      * |       144
#>                                     signalValue    pValue    qValue      peak
#>                                       <numeric> <numeric> <numeric> <numeric>
#>    T47D_ER_E2_2_mapped_reads_peak_1     7.42145  12.78537   9.77580       165
#>    T47D_ER_E2_2_mapped_reads_peak_2   141.93524 610.62402 605.03973       605
#>    T47D_ER_E2_2_mapped_reads_peak_3    28.75812  82.09051  78.39533       179
#>    T47D_ER_E2_2_mapped_reads_peak_4    33.39653  99.75146  95.96765       327
#>    T47D_ER_E2_2_mapped_reads_peak_5     5.89933   9.56727   6.65018       135
#>    T47D_ER_E2_2_mapped_reads_peak_6     5.56609   8.38709   5.51349       139
#>    T47D_ER_E2_2_mapped_reads_peak_7    14.37906  32.45039  29.12307       395
#>    T47D_ER_E2_2_mapped_reads_peak_8    25.04739  68.45815  64.84392       145
#>    T47D_ER_E2_2_mapped_reads_peak_9     4.63841   6.38478   3.59695       120
#>   T47D_ER_E2_2_mapped_reads_peak_10     9.27681  17.60511  14.49095       136
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#>