Make a set of consensus peaks

Make a set of consensus peaks based on umber of replicates

makeConsensus(
  x,
  p = 0,
  var = NULL,
  ignore.strand = TRUE,
  simplify = FALSE,
  ...
)

Arguments

x

A GRangesList

p

The minimum proportion of samples (i.e. elements of x) required for a peak to be retained in the output. By default all merged peaks will be returned

var

Additional columns in the mcols element to retain

ignore.strand, simplify, ...

Passed to reduceMC()

Value

GRanges object with mcols containing a logical vector for every element of x, along with the column n which adds all logical columns.

If any additional columns have been requested using var, these will be returned as CompressedList objects as produced by reduceMC().

Details

This takes a list of GRanges objects and forms a set of consensus peaks using the minimum proportion of replicates specified

See also

Examples

a <- GRanges("chr1:11-20")
a$score <- 1
b <- GRanges(c("chr1:18-22", "chr1:1-5"))
b$score <- c(0.6, 0.3)
grl <- GRangesList(a = a, b = b)
makeConsensus(grl)
#> GRanges object with 2 ranges and 3 metadata columns:
#>       seqnames    ranges strand |         a         b         n
#>          <Rle> <IRanges>  <Rle> | <logical> <logical> <numeric>
#>   [1]     chr1       1-5      * |     FALSE      TRUE         1
#>   [2]     chr1     11-22      * |      TRUE      TRUE         2
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
makeConsensus(grl, p = 1)
#> GRanges object with 1 range and 3 metadata columns:
#>       seqnames    ranges strand |         a         b         n
#>          <Rle> <IRanges>  <Rle> | <logical> <logical> <numeric>
#>   [1]     chr1     11-22      * |      TRUE      TRUE         2
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
makeConsensus(grl, p = 1, var = "score")
#> GRanges object with 1 range and 4 metadata columns:
#>       seqnames    ranges strand |         score         a         b         n
#>          <Rle> <IRanges>  <Rle> | <NumericList> <logical> <logical> <numeric>
#>   [1]     chr1     11-22      * |       1.0,0.6      TRUE      TRUE         2
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths