Map Genomic Ranges to genes using defined features

Map Genomic Ranges to genes using defined regulatory features

mapByFeature(
  gr,
  genes,
  prom,
  enh,
  gi,
  cols = c("gene_id", "gene_name", "symbol"),
  gr2prom = 0,
  gr2enh = 0,
  gr2gi = 0,
  gr2gene = 1e+05,
  prom2gene = 0,
  enh2gene = 1e+05,
  gi2gene = 0,
  ...
)

Arguments

gr
GRanges object with query ranges to be mapped to genes

genes
GRanges object containing genes (or any other nominal feature) to be assigned

prom
GRanges object defining promoters

enh
GRanges object defining Enhancers

gi
GInteractions object defining interactions. Mappings from interactions to genes should be performed as a separate prior step.

cols
Column names to be assigned as mcols in the output. Columns must be minimally present in genes. If all requested colunms are found in any of prom, enh or gi, these pre-existing mappings will be preferentially used. Any columns not found in utilised reference objects will be ignored.

gr2prom
The maximum permissible distance between a query range and any ranges defined as promoters

gr2enh
The maximum permissible distance between a query range and any ranges defined as enhancers

gr2gi
The maximum permissible distance between a query range and any ranges defined as GInteraction anchors

gr2gene
The maximum permissible distance between a query range and genes (for ranges not otherwise mapped)

prom2gene
The maximum permissible distance between a range provided in prom and a gene

enh2gene
The maximum permissible distance between a range provided in enh and a gene

gi2gene
The maximum permissible distance between a GInteractions anchor (provided in gi) and a gene

...
Passed to findOverlaps nad overlapsAny internally

Value

A GRanges object with added mcols as specified

Details

This function is able to utilise feature-level information and long-range interactions to enable better mapping of regions to genes. If provided, this essentially maps from ranges to genes using the regulatory features as a framework. The following sequential strategy is used:

Ranges overlapping a promoter are assigned to that gene

Ranges overlapping an enhancer are assigned to all genes within a specified distance

Ranges overlapping a long-range interaction are assigned to all genes connected by the interaction

Ranges with no gene assignment from the previous steps are assigned to all genes within a specified distance

If information is missing for one of these steps, the algorithm will simply proceed to the next step. If no promoter, enhancer or interaction data is provided, all ranges will be simply mapped by step 4.

Distances between each set of features and the query range can be individually specified by modifying the gr2prom, gr2enh, gr2gi or gr2gene parameters. Distances between features and genes can also be set using the parameters prom2gene, enh2gene and gi2gene.

Additionally, if previously defined mappings are included with any of the prom, enh or gi objects, this will be used in preference to any obtained from the genes object.

Examples

## Define some genes
genes <- GRanges(c("chr1:2-10:*", "chr1:25-30:-", "chr1:31-40:+"))
genes$gene_id <- paste0("gene", seq_along(genes))
## Add a promoter for each gene
prom <- promoters(genes, upstream = 1, downstream = 1)
## Some ranges to map
gr <- GRanges(paste0("chr1:", seq(0, 50, by = 10)))
mapByFeature(gr, genes)

#> GRanges object with 6 ranges and 1 metadata column:
#>       seqnames    ranges strand |           gene_id
#>          <Rle> <IRanges>  <Rle> |   <CharacterList>
#>   [1]     chr1         0      * | gene1,gene2,gene3
#>   [2]     chr1        10      * | gene1,gene2,gene3
#>   [3]     chr1        20      * | gene1,gene2,gene3
#>   [4]     chr1        30      * | gene1,gene2,gene3
#>   [5]     chr1        40      * | gene1,gene2,gene3
#>   [6]     chr1        50      * | gene1,gene2,gene3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

mapByFeature(gr, genes, prom)

#> GRanges object with 6 ranges and 1 metadata column:
#>       seqnames    ranges strand |           gene_id
#>          <Rle> <IRanges>  <Rle> |   <CharacterList>
#>   [1]     chr1         0      * |             gene1
#>   [2]     chr1        10      * | gene1,gene2,gene3
#>   [3]     chr1        20      * | gene1,gene2,gene3
#>   [4]     chr1        30      * |       gene2,gene3
#>   [5]     chr1        40      * | gene1,gene2,gene3
#>   [6]     chr1        50      * | gene1,gene2,gene3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths