Merge sliding windows using a specified column

Merge sliding windows using a specified column

mergeByCol(x, ...)

# S4 method for GenomicRanges
mergeByCol(
  x,
  df = NULL,
  col,
  by = c("max", "median", "mean", "min"),
  logfc = "logFC",
  pval = "P",
  inc_cols,
  p_adj_method = "fdr",
  merge_within = 1L,
  ignore_strand = FALSE,
  ...
)

# S4 method for RangedSummarizedExperiment
mergeByCol(
  x,
  df = NULL,
  col,
  by = c("max", "median", "mean", "min"),
  logfc = "logFC",
  pval = "P",
  inc_cols,
  p_adj_method = "fdr",
  merge_within = 1L,
  ignore_strand = FALSE,
  ...
)

Arguments

x
A GenomicRanges or SummarizedExperiment object

...
Not used

df
A data.frame-like object containing the columns of interest. If not provided, any columns in the mcols() slot will be used.

col
The column to select as representative of the merged ranges

by
The method for selecting representative values

logfc
Column containing logFC values

pval
Column containing p-values

inc_cols
Any additional columns to return. Output will always include columns specified in the arguments col, logfc and pval. Note that values from any additional columns will correspond to the selected range returned in keyval_range

p_adj_method
Any of p.adjust.methods

merge_within
Merge any ranges within this distance

ignore_strand
Passed internally to reduce and findOverlaps

Value

A Genomic Ranges object

Details

This merges sliding windows by a given column, returning representative values from each requested column, corresponding to the selected row. Values from the requested column can be selected using any of min(), max(), mean() or median(), although max() is strongly recommended. The range corresponding to the returned values is returned as the column. keyval_range.

Taking the p-value from the selected row, the number of windows with lower p-values are counted by direction and returned in the final object

Examples

x <- GRanges(c("chr1:1-10", "chr1:6-15", "chr1:51-60"))
df <- data.frame(logFC = rnorm(3), logCPM = rnorm(3,8), p = 10^-rexp(3))
mergeByCol(x, df, col = "logCPM", pval = "p")
#> GRanges object with 2 ranges and 8 metadata columns:
#>       seqnames    ranges strand | n_windows      n_up    n_down keyval_range
#>          <Rle> <IRanges>  <Rle> | <integer> <integer> <integer>    <GRanges>
#>   [1]     chr1      1-15      * |         2         0         2    chr1:6-15
#>   [2]     chr1     51-60      * |         1         0         1   chr1:51-60
#>          logCPM     logFC           p        fdr
#>       <numeric> <numeric>   <numeric>  <numeric>
#>   [1]  10.06502 -0.282705 0.419650740 0.41965074
#>   [2]   6.36901 -0.553699 0.000238735 0.00047747
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths