Detect Differential ChIP Signal

Detect differential ChIP signal using one of many approaches

fitAssayDiff(x, ...)

# S4 method for SummarizedExperiment
fitAssayDiff(
  x,
  assay = "counts",
  design = NULL,
  coef = NULL,
  lib.size = "totals",
  method = c("qlf", "lt"),
  norm = c("none", "TMM", "RLE", "TMMwsp", "upperquartile"),
  groups = NULL,
  fc = 1,
  lfc = log2(fc),
  asRanges = FALSE,
  offset = NULL,
  null = c("interval", "worst.case"),
  weighted = FALSE,
  ...
)

Arguments

x

a SummarizedExperiment object

...

Passed to calcNormFactors and glmQLFit when method = "qlf". If method = "lt", instead passed to lmFit, treat, eBayes

assay

The assay to use for analysis

design

The design matrix to use for analysis

coef

The required column from the design matrix

lib.size

The column within the colData element which contains the library size information. If set to NULL, column summaries will be used.

method

the analytic method to be used. Can be 'qlf' which will fit counts using the glmQLFit strategy , or 'lt' which fits the limma-trend model on logCPM, or pre-processed logCPM values

norm

The normalisation strategy to use when running the glmQLF models. The value 'none' relies solely on library-size normalisation, and is the default. All methods available in calcNormFactors are implemented. Ignored when using method = "lt"

groups

character(1) If a column name is supplied here, group-based normalisation will be applied to GLM models treating data in this column as a grouping factor. Ignored when using method = "lt"

fc, lfc

Thresholds passed to treat or glmTreat

asRanges

logical(1). By default, the returned object will be a SummarizedExperiment object with the results added to the rowData element. Setting asRanges = TRUE will only return the GRanges object from this element

offset

If provided will be used as the offset when the DGEList object is created during model fitting

null

Passed to glmTreat

weighted

logical(1) Passed to calcNormFactors

Value

A SummarizedExperiment object with results set as the rowData element. Any existing columns not contained in the differential ChIP results will be retained. Results from testing will contain logCPM, logFC, PValue and the t/F statistic as appropriate, along with an FDR-adjusted p-value

Details

Starting with a SummarizedExperiment object this function fits either a glmQLFit model to count data, or the limma-trend model to logCPM data.

If fitting Generalised Linear Models via glmQLFit, options for normalisation are "none", which normalises to library size. Existing library sizes are commonly found in the "totals" column of the colData element and this is attempted by default. All methods provided in calcNormFactors are also implemented, with the added possibility of normalising within groups instead of across the entire dataset. To enable this, the column with the grouping factor is expected to be in the colData element and is simply called by column name. No normalisation is applied when using the limma-trend model, as this allows for previous normalisation strategies to be performed on the data.

Normalising to ChIP Input samples, or using offsets is not yet implemented.

Either range-based hypothesis testing is implemented using glmTreat or treat. Setting fc to 1 (or lfc to 0) will default to a point-based null hypothesis, equivalent to either glmQLFTest (method = "qlf") or eBayes (method = "lt").

It should also be noted that this is primarily a convenience function and if requiring intermediate output from any setps, then these can be run individually as conventionally specified.

Examples

nrows <- 200; ncols <- 6
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
colnames(counts) <- paste0("Sample_", seq_len(ncols))
df <- DataFrame(treat = c("A", "A", "A", "B", "B", "B"))
df$treat <- as.factor(df$treat)
se <- SummarizedExperiment(
  assays = SimpleList(counts = counts), colData = df
)
X <- model.matrix(~treat, colData(se))
se <- fitAssayDiff(se, design = X, lib.size = NULL)
#> Creating DGE list...
#> Calculating experiment-wide normalisation factors...
#> Estimating dispersions...
#> Running glmQLFit...
rowData(se)
#> DataFrame with 200 rows and 5 columns
#>          logFC    logCPM           F     PValue       FDR
#>      <numeric> <numeric>   <numeric>  <numeric> <numeric>
#> 1    4.3066297   11.7021 10.16802466 0.00399995  0.799990
#> 2    0.0834144   12.2407  0.00642621 0.93678278  0.996477
#> 3    1.6795006   11.6748  1.92608248 0.17814742  0.996477
#> 4   -1.1910934   12.0524  1.30459678 0.26482720  0.996477
#> 5    0.5179778   12.0442  0.23548266 0.63196061  0.996477
#> ...        ...       ...         ...        ...       ...
#> 196  0.0907519   12.7687  0.02703171  0.8708054  0.996477
#> 197  1.0952799   12.3006  1.13433141  0.2976250  0.996477
#> 198 -1.1430288   12.6072  2.94270952  0.0993631  0.996477
#> 199  0.1461143   12.3253  0.02388492  0.8784917  0.996477
#> 200  0.0463691   11.8591  0.00110078  0.9738116  0.996477