Get Profile Data surrounding specified ranges

Get coverage Profile Data surrounding specified ranges

getProfileData(x, gr, ...)

# S4 method for BigWigFile,GenomicRanges
getProfileData(
  x,
  gr,
  upstream = 2500,
  downstream = upstream,
  bins = 100,
  mean_mode = "w0",
  ...
)

# S4 method for BigWigFileList,GenomicRanges
getProfileData(
  x,
  gr,
  upstream = 2500,
  downstream = upstream,
  bins = 100,
  mean_mode = "w0",
  ...
)

# S4 method for character,GenomicRanges
getProfileData(
  x,
  gr,
  upstream = 2500,
  downstream = upstream,
  bins = 100,
  mean_mode = "w0",
  ...
)

# S4 method for character,GenomicRanges
getProfileData(
  x,
  gr,
  upstream = 2500,
  downstream = upstream,
  bins = 100,
  mean_mode = "w0",
  ...
)

# S4 method for ANY,ANY
getProfileData(x, gr)

Arguments

x

A BigWigFile or BigWiFileList

gr

A GRanges object

...

Passed to normalizeToMatrix

upstream

The distance to extend upstream from the centre of each range within gr

downstream

The distance to extend downstream from the centre of each range within gr

bins

The total number of bins to break the extended ranges into

mean_mode

The method used for calculating the score for each bin. See normalizeToMatrix for details

Value

GRanges or GrangesList with column profile_data, as described above

Details

This will take all provided ranges and set as identical width ranges, extending by the specified amount both up and downstream of the centre of the provided ranges. By default, the ranges extensions are symmetrical and only the upstream range needs to be specified, however this parameterisation allows for non-symmetrical ranges to be generated.

These uniform width ranges will then be used to extract the value contained in the score field from one or more BigWigFiles. Uniform width ranges are then broken into bins of equal width and the average score found within each bin.

The binned profiles are returned as a DataFrameList called profile_data as a column within the resized GRanges object. Column names in each DataFrame are score, position and bp.

Examples

bw <- system.file("tests", "test.bw", package = "rtracklayer")
gr <- GRanges("chr2:500")
pd <- getProfileData(bw, gr, upstream = 100, bins = 10)
pd
#> GRanges object with 1 range and 1 metadata column:
#>       seqnames    ranges strand |                           profile_data
#>          <Rle> <IRanges>  <Rle> |                   <SplitDataFrameList>
#>   [1]     chr2   400-599      * | -0.75:u1:-90,-0.75:u2:-70,-0.75:u3:-50
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
pd$profile_data
#> SplitDataFrameList of length 1
#> [[1]]
#> DataFrame with 10 rows and 3 columns
#>        score position        bp
#>    <numeric> <factor> <numeric>
#> 1      -0.75       u1       -90
#> 2      -0.75       u2       -70
#> 3      -0.75       u3       -50
#> 4      -0.75       u4       -30
#> 5      -0.75       u5       -10
#> 6      -0.75       d1        10
#> 7      -0.75       d2        30
#> 8      -0.75       d3        50
#> 9      -0.75       d4        70
#> 10     -0.75       d5        90
#>